refactor(common): move chain value types into chain_config module
